switchboard: suppress handshake failure messages
the process_request handler is unable to handle invalid
HTTP requests (such as HEAD and POST), so this is a workaround
to supress upstream handshake failure error messages.

ideally we would front this service with a http proxy to make the
requests safe / provide authentication / &c &c and not have to
ignore most websockets.server logs.